+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 * Change `--bias` to be optional on `px master flat` and allow the bias to be found in the index
 * `px obs calibrate` no longer enforces bias, dark, or flat to support smart scopes that just need debayering
 * Rework project config and implement `single` and `spiral_mosiac` stacking for projects
+* Changed `px project init` to be interactive and build a project config.
 
 ## 0.1.3